Win32 you idiot

+++ Vorbisfile.cs
@@ -77,7 +77,7 @@ public static class Vorbisfile
 	[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 	public delegate int seek_func(
 		IntPtr datasource,	// Refers to a void*
-		long offset,
+		IntPtr offset, // Refers to a long, ignoring Win64!
 		SeekWhence whence
 	);
 
@@ -86,8 +86,9 @@ public static class Vorbisfile
 		IntPtr datasource	// Refers to a void*
 	);
 
+	/* IntPtr refers to a long, ignoring Win64! */
 	[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-	public delegate long tell_func(
+	public delegate IntPtr tell_func(
 		IntPtr datasource	// Refers to a void*
 	);
 
@@ -100,12 +101,12 @@ public static class Vorbisfile
 	{
 		public int version;
 		public int channels;
-		public long rate;
-		public long bitrate_upper;
-		public long bitrate_nominal;
-		public long bitrate_lower;
-		public long bitrate_window;
-		public IntPtr codec_setup; // Refers to a void*
+		public IntPtr rate;		// refers to a long, ignoring Win64!
+		public IntPtr bitrate_upper;	// refers to a long, ignoring Win64!
+		public IntPtr bitrate_nominal;	// refers to a long, ignoring Win64!
+		public IntPtr bitrate_lower;	// refers to a long, ignoring Win64!
+		public IntPtr bitrate_window;	// refers to a long, ignoring Win64!
+		public IntPtr codec_setup;	// Refers to a void*
 	}
 
 	[StructLayout(LayoutKind.Sequential)]
@@ -204,8 +205,9 @@ public static class Vorbisfile
 	[DllImport(nativeLibName, CallingConvention = CallingConvention.Cdecl)]
 	public static extern double ov_time_total(IntPtr vf, int i);
 
+	/* IntPtr refers to a long, ignoring Win64! */
 	[DllImport(nativeLibName, CallingConvention = CallingConvention.Cdecl)]
-	public static extern long ov_read(
+	public static extern IntPtr ov_read(
 		IntPtr vf,
 		byte[] buffer,
 		int length,
@@ -215,8 +217,9 @@ public static class Vorbisfile
 		out int current_section
 	);
 
+	/* IntPtr refers to a long, ignoring Win64! */
 	[DllImport(nativeLibName, CallingConvention = CallingConvention.Cdecl)]
-	public static extern long ov_read(
+	public static extern IntPtr ov_read(
 		IntPtr vf,
 		IntPtr buffer,
 		int length,
